Clean up コンポーネント分割
