+++ pages/todo/_id/index.vue
@@ -1,23 +1,6 @@
 <template>
   <section v-if="!isLoading">
-    <h1 class="text-3xl break-all">
-      Todo No.{{ currentId }} {{ currentTodo.title }}
-    </h1>
-    <p class="text-xs mt-2">
-      <span
-        v-if="currentTodo.isDone"
-        class="inline-block text-green-600 border border-green-600 px-4 py-1"
-      >
-        完了
-      </span>
-      <span
-        v-else
-        class="inline-block text-blue-600 border border-blue-600 px-4 py-1"
-      >
-        作業中
-      </span>
-    </p>
-    <p class="mt-4 p-4 border bg-gray-100 break-all">{{ currentTodo.text }}</p>
+    <TodoDetail />
     <div class="text-right mt-6">
       <NuxtLink class="text-sm text-blue-600" to="/">トップへ戻る</NuxtLink>
     </div>
@@ -26,15 +9,17 @@
 
 <script>
 import { mapGetters, mapMutations } from 'vuex'
+import TodoDetail from '~/components/TodoDetail'
 
 export default {
   name: 'Todo',
   layout: 'todo',
+  components: {
+    TodoDetail,
+  },
   computed: {
     ...mapGetters({
       isLoading: 'loading/isLoading',
-      currentId: 'todos/currentId',
-      currentTodo: 'todos/currentTodo',
       allTodos: 'todos/allTodos',
     }),
   },
@@ -43,7 +28,6 @@ export default {
     if (!this.allTodos.length) {
       return this.$router.push('/')
     }
-    this.setCurrentId(this.$route.params.id)
     this.changeLoading(false)
   },
   destroyed() {
@@ -52,7 +36,6 @@ export default {
   methods: {
     ...mapMutations({
       changeLoading: 'loading/changeLoading',
-      setCurrentId: 'todos/setCurrentId',
     }),
   },
 }
