+++ store/todos.js
@@ -1,5 +1,6 @@
 export const state = () => ({
   list: [], // ToDOリスト
+  currentFilter: 'all', // 現在のフィルタリングの状態
   currentId: 0, // 現在のページのToDoのID
 })
 
@@ -12,6 +13,30 @@ export const getters = {
    */
   allTodos(state) {
     return state.list
+  },
+  /**
+   * フィルタリングされたToDoを返す
+   *
+   * @param {Object} state
+   * @return {Array}
+   */
+  filteredTodos(state) {
+    if (state.currentFilter === 'working') {
+      return state.list.filter((todo) => !todo.isDone)
+    } else if (state.currentFilter === 'done') {
+      return state.list.filter((todo) => todo.isDone)
+    } else {
+      return state.list
+    }
+  },
+  /**
+   * 現在のページ以外のToDoを返す
+   *
+   * @param {Object} state
+   * @return {Array}
+   */
+  otherTodos(state) {
+    return state.list.filter((todo) => todo.id !== state.currentId)
   },
   /**
    * 現在のページのToDoのIDを返す
@@ -32,13 +57,13 @@ export const getters = {
     return state.list.filter((todo) => todo.id === state.currentId)[0]
   },
   /**
-   * 現在のページ以外のToDoを返す
+   * 現在のフィルタリングの状態を返す
    *
    * @param {Object} state
-   * @return {Array}
+   * @return {String}
    */
-  otherTodos(state) {
-    return state.list.filter((todo) => todo.id !== state.currentId)
+  currentFilter(state) {
+    return state.currentFilter
   },
 }
 
@@ -72,13 +97,22 @@ export const mutations = {
    * @param {Object} state
    * @param {Number} id
    */
-  toggleTodo(state, id) {
+  toggleTodoState(state, id) {
     state.list = state.list.map((todo) => {
       if (todo.id === id) {
         todo.isDone = !todo.isDone
       }
       return todo
     })
+  },
+  /**
+   * filterの状態を変更する
+   *
+   * @param {Object} state
+   * @param {String} filter
+   */
+  setCurrentFilter(state, filter) {
+    state.currentFilter = filter
   },
   /**
    * 現在のページのToDoのIDをセットする
